Add player team data
Add the player's teams wins, losses, city, name and key data to the player's teams array for calculating wins.
